requirements.txt
aiohttp==3.4.4
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3.1
future==0.16.0
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.4.0
pycparser==2.18
pylint==2.1.1
python-telegram-bot==11.1.0
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
yarl==1.2.6
